api/requirements.txt
Django==2.0.1
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
Pillow
requests
django-cors-headers

markdown
html2text
pycurl
bleach
beautifulsoup4
geocoder
numpy